+++ packages/apollo-link-batch-http/src/batchHttpLink.ts
@@ -1,21 +1,18 @@
 import { ApolloLink, Operation, FetchResult, Observable } from 'apollo-link';
-import { ApolloFetch, createApolloFetch } from 'apollo-fetch';
+import {
+  serializeFetchBody,
+  selectURI,
+  parseAndCheckHttpResponse,
+  checkFetcher,
+  selectHttpOptionsAndBody,
+  createSignalIfSupported,
+  fallbackHttpConfig,
+  HttpOptions,
+} from 'apollo-link-http-common';
 import { BatchLink } from 'apollo-link-batch';
 
-import { print } from 'graphql/language/printer';
-
 export namespace BatchHttpLink {
-  /**
-   * Note: This package will be updated to remove the dependency on apollo-fetch an use the same options / API as the http-link.
-   */
-  export interface Options {
-    /**
-     * The URI to use when fetching operations.
-     *
-     * Defaults to '/graphql'.
-     */
-    uri?: string;
-
+  export interface Options extends HttpOptions {
     /**
      * The maximum number of operations to include in one fetch.
      *
@@ -31,65 +28,140 @@ export namespace BatchHttpLink {
     batchInterval?: number;
 
     /**
-     * An instance of ApolloFetch to use when making requests.
+     * Sets the key for an Operation, which specifies the batch an operation is included in
      */
-    fetch?: ApolloFetch;
+    batchKey?: (Operation) => string;
   }
 }
 
-/** Transforms Operation for into HTTP results.
+/**
+ * Transforms Operation for into HTTP results.
  * context can include the headers property, which will be passed to the fetch function
  */
 export class BatchHttpLink extends ApolloLink {
-  private headers = {};
-  private apolloFetch: ApolloFetch;
   private batchInterval: number;
   private batchMax: number;
   private batcher: ApolloLink;
 
-  constructor(fetchParams?: BatchHttpLink.Options) {
+  constructor(fetchParams: BatchHttpLink.Options = {}) {
     super();
 
-    this.batchInterval = (fetchParams && fetchParams.batchInterval) || 10;
-    this.batchMax = (fetchParams && fetchParams.batchMax) || 10;
+    let {
+      uri = '/graphql',
+      // use default global fetch is nothing passed in
+      fetch: fetcher,
+      includeExtensions,
+      batchInterval,
+      batchMax,
+      batchKey,
+      ...requestOptions
+    } = fetchParams;
 
-    this.apolloFetch =
-      (fetchParams && fetchParams.fetch) ||
-      createApolloFetch({ uri: fetchParams && fetchParams.uri });
+    // dev warnings to ensure fetch is present
+    checkFetcher(fetcher);
 
-    this.apolloFetch.batchUse((request, next) => {
-      request.options.headers = {
-        ...request.options.headers,
-        ...this.headers,
-      };
-      next();
-    });
+    //fetcher is set here rather than the destructuring to ensure fetch is
+    //declared before referencing it. Reference in the destructuring would cause
+    //a ReferenceError
+    if (!fetcher) {
+      fetcher = fetch;
+    }
+
+    const linkConfig = {
+      http: { includeExtensions },
+      options: requestOptions.fetchOptions,
+      credentials: requestOptions.credentials,
+      headers: requestOptions.headers,
+    };
+
+    this.batchInterval = batchInterval || 10;
+    this.batchMax = batchMax || 10;
 
     const batchHandler = (operations: Operation[]) => {
-      return new Observable<FetchResult[]>(observer => {
-        const printedOperations = operations.map((operation: Operation) => ({
-          ...operation,
-          query: print(operation.query),
-        }));
+      const chosenURI = selectURI(operations[0], uri);
 
-        this.apolloFetch(printedOperations)
-          .then(data => {
-            observer.next(data);
+      const context = operations[0].getContext();
+
+      const contextConfig = {
+        http: context.http,
+        options: context.fetchOptions,
+        credentials: context.credentials,
+        headers: context.headers,
+      };
+
+      //uses fallback, link, and then context to build options
+      const optsAndBody = operations.map(operation =>
+        selectHttpOptionsAndBody(
+          operation,
+          fallbackHttpConfig,
+          linkConfig,
+          contextConfig,
+        ),
+      );
+
+      const body = optsAndBody.map(({ body }) => body);
+      const options = optsAndBody[0].options;
+
+      try {
+        (options as any).body = serializeFetchBody(body);
+      } catch (parseError) {
+        return new Observable<FetchResult[]>(observer => {
+          observer.error(parseError);
+        });
+      }
+
+      const { controller, signal } = createSignalIfSupported();
+      if (controller) (options as any).signal = signal;
+
+      return new Observable<FetchResult[]>(observer => {
+        // the raw response is attached to the context in the BatchingLink
+        fetcher(chosenURI, options)
+          .then(parseAndCheckHttpResponse(operations))
+          .then(result => {
+            // we have data and can send it to back up the link chain
+            observer.next(result);
             observer.complete();
+            return result;
           })
-          .catch(observer.error.bind(observer));
+          .catch(err => {
+            // fetch was cancelled so its already been cleaned up in the unsubscribe
+            if (err.name === 'AbortError') return;
+            observer.error(err);
+          });
+
+        return () => {
+          // XXX support canceling this request
+          // https://developers.google.com/web/updates/2017/09/abortable-fetch
+          if (controller) controller.abort();
+        };
       });
     };
+
+    batchKey =
+      batchKey ||
+      ((operation: Operation) => {
+        const context = operation.getContext();
+
+        const contextConfig = {
+          http: context.http,
+          options: context.fetchOptions,
+          credentials: context.credentials,
+          headers: context.headers,
+        };
+
+        //may throw error if config not serializable
+        return selectURI(operation, uri) + JSON.stringify(contextConfig);
+      });
 
     this.batcher = new BatchLink({
       batchInterval: this.batchInterval,
       batchMax: this.batchMax,
+      batchKey,
       batchHandler,
     });
   }
 
   public request(operation: Operation): Observable<FetchResult> | null {
-    this.headers = operation.getContext().headers || this.headers;
     return this.batcher.request(operation);
   }
 }