+++ rollup.config.js
@@ -4,6 +4,12 @@ export const globals = {
   'apollo-link': 'apolloLink.core',
   'apollo-link-batch': 'apolloLink.batch',
   'apollo-utilities': 'apollo.utilities',
+
+  //GraphQL
+  'graphql/language/printer': 'printer',
+
+  //zen-observable
+  'zen-observable': 'Observable',
 };
 
 export default name => ({