+++ packages/apollo-link-batch-http/CHANGELOG.md
@@ -1,5 +1,6 @@
 ### vNext
-- deprecate package in favor of batching in apollo-link-http [PR#364](https://github.com/apollographql/apollo-link/pull/364)
+- share logic with apollo-link-http through apollo-link-http-core [PR#364](https://github.com/apollographql/apollo-link/pull/364)
+- remove apollo-fetch [PR#364](https://github.com/apollographql/apollo-link/pull/364)
 
 ### 1.0.5
 - ApolloLink upgrade